services/api/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
SQLAlchemy==2.0.35
httpx==0.27.2
python-jose==3.3.0
psycopg[binary]==3.2.1
